Aggiunta funzionalità per la gestione delle variabili di tessuto e colore del prodotto
- Implementato il metodo `ensureTessuPlaceholderWithMeters` nel modello Prodotto per gestire i requisiti di tessuto.
- Aggiunta migrazione per modificare il tipo di campo `quantity` nella tabella `product_components` in decimale.
- Aggiornata la finestra modale di creazione prodotto per includere l'input per i metri di tessuto richiesti.
- Migliorate le visualizzazioni dei dati master per visualizzare messaggi quando non vengono trovate categorie, componenti, clienti o fornitori.
- Introdotta la finestra modale delle variabili di prodotto per la gestione delle whitelist e delle sostituzioni di tessuti/colori.
- Aggiunti percorsi per la gestione delle variabili di prodotto e delle sostituzioni.

+++ database/migrations/2025_06_14_123716_create_product_components_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
         Schema::create('product_components', function (Blueprint $table) {
             $table->foreignId('product_id')->constrained()->cascadeOnDelete();
             $table->foreignId('component_id')->constrained()->cascadeOnDelete();
-            $table->integer('quantity')->comment('Numero di componenti per prodotto');
+            $table->decimal('quantity', 8, 3)->comment('Numero di componenti per prodotto');
             $table->primary(['product_id','component_id']);
         });
     }
